refactor: 리포트 ID 상태 없을 때 일회성 조회 커스텀 훅 분리
