+++ frontend/src/hooks/report/usePollReportStatus.ts
@@ -28,18 +28,15 @@ export const areAllTasksTerminal = (status: ReportStatus): boolean => {
 }
 
 /**
- * 리포트의 상세 상태를 주기적으로 폴링하는 커스텀 훅.
- * 모든 하위 작업이 완료/실패 시 폴링을 자동으로 중지합니다.
+ * 리포트의 상세 상태를 주기적으로 폴링하는 커스텀 훅
+ * 모든 하위 작업이 완료/실패 시 폴링을 자동으로 중지
  * @param reportId - 조회할 리포트의 ID (number 타입)
  * @param options - 폴링 간격 등 추가 옵션
  */
 export const usePollReportStatus = (reportId: number | undefined, options: UseReportStatusOptions = {}) => {
     const { intervalMs = 3000, enabled = true } = options
 
-    // const user = useAuthStore((state) => state.user)
-    // const channelId = user?.channelId
     const channelId = useAuthStore((state) => state.user?.channelId)
-
     const { updateReportStatus, removeReportStatus, removePendingReportId, beginReportCleanup } = useReportStore(
         (state) => state.actions
     )
@@ -71,24 +68,15 @@ export const usePollReportStatus = (reportId: number | undefined, options: UseRe
         if (query.data && typeof reportId === 'number') {
             updateReportStatus(reportId, query.data)
 
-            if (areAllTasksTerminal(query.data)) {
-                removePendingReportId(reportId)
-            }
-
             const isAnyFailed = Object.values(query.data).includes('FAILED')
             if (isAnyFailed && !cleanupReportIds.includes(reportId)) {
-                // 정리 절차를 시작하고, 모든 관련 상태를 즉시 정리합니다.
                 beginReportCleanup(reportId)
-                deleteReport({ reportId })
                 removeReportStatus(reportId)
-                removePendingReportId(reportId) // 실패했으므로 폴링을 즉시 중단합니다.
-
-                // 실패 처리가 끝났으므로 더 이상 다른 조건을 확인할 필요가 없습니다.
+                removePendingReportId(reportId)
+                deleteReport({ reportId })
                 return
             }
 
-            // ✅ 2. 실패가 아닐 경우에만, 성공적으로 모든 작업이 완료되었는지 확인합니다.
-            // (isAnyFailed가 false이므로, 이 조건은 모든 상태가 'COMPLETED'일 때만 충족됩니다.)
             if (areAllTasksTerminal(query.data)) {
                 removePendingReportId(reportId)
             }
@@ -105,4 +93,31 @@ export const usePollReportStatus = (reportId: number | undefined, options: UseRe
     ])
 
     return query
+}
+
+/**
+ * 리포트의 상태를 일회성으로 조회해 스토어에 업데이트하는 커스텀 훅
+ * @param reportId - 조회할 리포트의 ID
+ * @returns isInvalidReportError - 초기 상태 조회 시 존재하지 않는 리포트 등의 에러 발생 여부
+ */
+export const useGetInitialReportStatus = (reportId: number) => {
+    const currentReportStatus = useReportStore((state) => state.statuses[reportId])
+    const updateReportStatus = useReportStore((state) => state.actions.updateReportStatus)
+
+    const { data: initialStatusData, isError: isInvalidReportError } = useQuery({
+        queryKey: ['reportStatus', reportId, 'initialCheck'],
+        queryFn: () => getReportStatus({ reportId }),
+        enabled: !!reportId && !currentReportStatus,
+        retry: false,
+        refetchOnWindowFocus: false,
+        select: (data) => data.result,
+    })
+
+    useEffect(() => {
+        if (initialStatusData) {
+            updateReportStatus(reportId, initialStatusData)
+        }
+    }, [initialStatusData, reportId, updateReportStatus])
+
+    return { isInvalidReportError }
 }
